feat(auth): implement oidc authentication
Implement OIDC-based authentication using `mozilla-django-oidc`.
Additionally to the default ModelBackend (used for django-admin),
we subclass the Mozilla OIDC backend and customize it for our needs.
That is, we accept two kinds of OIDC flows:
 * authorization code (using a public client)
 * client credentials grant (using a confidential client).

The first is the accepted for the timed-frontend, the latter is used for
server-to-server communication by services consuming the timed api.