chore: ember 4.4 upgrade (#1067)
* chore: upgrade to ember 4.4

BREAKING CHANGE: Requires node >= 16.

* chore: remove custom unique-id helper

It's included now in ember by default.

* chore(deps): use optional features package for jquery

* use optional-features package
* remove route-action-helper addon

* fix: replace ember-scrollable with simplebar

* fix(index-reports-route): add missing notification service

* fix(ember-cli-nouislider): prevent global usage of Logger

The logger was exported of the global Ember object which is
no longer available as of ember v4. Therefor we may use the
forked version of the addon. Replace addon completely in the
near future or upgrade original repository.

* refactor(unverfified-reports): prevent global ember object usage

* refactor(worktime-balance-chart): to glimmer component

* fix(page-ability): safe property access

* refactor(worktime-balance-chart): adapt integration test

* fix(user-edit-credits): do not assert default query params